- Feat: BaseEntity 생성
