+++ src/main/java/com/sprint/mission/discodeit/entity/BaseEntity.java
@@ -1,0 +1,34 @@
+package com.sprint.mission.discodeit.entity;
+
+import java.util.UUID;
+
+public class BaseEntity {
+
+    private static final long serialVersionUID = 1L;
+    private UUID id;
+    private Long createdAt;
+    private Long updatedAt;
+
+    public BaseEntity() {
+        this.id = UUID.randomUUID();
+        this.createdAt = System.currentTimeMillis();
+        this.updatedAt = null;
+    }
+
+    public UUID getId() {
+        return id;
+    }
+
+    public Long getCreatedAt() {
+        return createdAt;
+    }
+
+    public Long getUpdatedAt() {
+        return updatedAt;
+    }
+
+    public void setUpdatedAt() {
+        this.updatedAt = System.currentTimeMillis();
+    }
+
+}
